Adjust chunk size of compressed headers for PUSH_PROMISE
When sending a PUSH_PROMISE that results in a CONTINUATION, we would end
up trying to send a PUSH_PROMISE that was actually
(MAX_HTTP_PAYLOAD_SIZE + 4) bytes long, since the PUSH_PROMISE uses 4
bytes of the payload for the promised stream ID. This fixes that by
cutting a PUSH_PROMISE into (MAX_HTTP_PAYLOAD - 4) byte chunks. That
means that any CONTINUATION frames from a PUSH_PROMISE will be
"underfilled" by 4 bytes, but that's a minor inefficiency we can live
with.

Fixes #93