Use a new agent for request-specific TLS settings.
If the settings are passed to `http2.Agent#request()`, and an `agent`
setting is not specified, a new backing `https.Agent` will be used for
that request. Otherwise, the default agent will be used. This matches
the Node `https.request()` behavior.

See https://nodejs.org/api/https.html#https_https_request_options_callback
for a list of TLS settings that can be overridden for each request.