Emit sentFrame and receivedFrame events from endpoint.
Want to be able to inspect frames as the come and go.  Added an onEndpoint parameter to Agent.request() and EndPoint().  Needed onEndpoint because after constructor for EndPoint has returned some frames have already been sent, and would be missed by any event handler.  Also added onSentFrame and onReceivedFrame parameters to Connection() constructor for same reason.

Refactored the constructor for EndPoint and Connection to take a config object instead of a list of parameters.  The parameter list on the constructor was getting pretty long.

Added test for sentFrame, and receiveFrame.  Updated existing tests to handle new constructor for EndPoint and Connection.

Conflicts:
	lib/http.js