Do not continue processing streams with malformed headers
When receiving a malformed HEADERS (missing one of the : headers), we
would send a RST_STREAM and then continue processing the stream as if
nothing had gone wrong. When the higher layer then asked us to send a
response on said stream, we would try to send our own HEADERS even
though the stream is CLOSED. We (righly) check to ensure that we weren't
sending HEADERS on a CLOSED stream, and would crash when we did. This
prevents us from getting to that point by early-returning if we don't
get all the : headers we need.

Fixes 145
Fixes 147