Fix #146
We were not properly sanity-checking frames when trying to pass them
onto a stream - so we would allow a GOAWAY to get to the handler for a
stream, for instance, which would cause crashiness.

This adds some sanity checks to ensure we receive frames on/off a stream
as appropriate, as well as to ensure we don't create a stream unless
it's from a HEADERS (which would also be illegal).