Roll back part of the "fix" for 146.
I was incorrect to prevent anything but HEADERS from creating a stream
within node-http2 - Firefox sends unused stream IDs to create a priority
stream. The HEADERS-only sanity check broke that (and is wrong - we have
multiple independent implementations that accept Firefox's behavior).

Note that this does not actually make 146 a problem again - this was
just an extra change that I slid in with the real fix for that issue.

Fixes #167